feat: Add `ServerCallContext` into task store operations (#443)
In production systems the persistence of the task should be protected
based on the credentials of the creator of the task (e.g. a user id or
email). Additionally, applications may have other criteria to use for
task persistence (like application name, or region task runs in).

---

Providing the `ServerCallContext` into the calls to the `get`, `save`
and `delete` interface for the task store allows customization of the
persisted task data based on the characteristics needed for a real
solution. Agent implementors can construct the appropriate
`ServerCallContext` based on the incoming request and use that
information at task creation, retrieval and deletion time.

Fixes #442  🦕

---------

Co-authored-by: Holt Skinner <13262395+holtskinner@users.noreply.github.com>
Co-authored-by: Holt Skinner <holtskinner@google.com>